% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tariff_1.0-package.R
\docType{data}
\name{causetext}
\alias{causetext}
\title{Translation list of COD codes}
\format{A data frame with the translation of COD codes to their names on 68
CODs (both the version of COD only and COD with group code).}
\description{
This is the translation of COD abbreviation codes into their corresponding
full names.
}
\examples{
data(causetext)
}
\keyword{datasets}

