% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tarrif.summary.r
\name{print.tariff_summary}
\alias{print.tariff_summary}
\title{Print method for the summary of the results obtained from Tariff algorithm}
\usage{
\method{print}{tariff_summary}(x, ...)
}
\arguments{
\item{x}{summary object for Tariff fit}

\item{...}{not used}
}
\description{
This function prints the summary message of the fitted results.
}
\keyword{Tariff}

