% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tariff_1.0-package.R
\docType{data}
\name{SampleCategory3}
\alias{SampleCategory3}
\title{Grouping of causes in RandomVA3}
\format{17 by 2 matrix}
\description{
This is a matrix specifying a default grouping of the causes used in RandomVA3.
}
\examples{

data(SampleCategory3)
SampleCategory3
}
\keyword{datasets}
