% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tariff.plot.r
\name{Tariff.csmf}
\alias{Tariff.csmf}
\title{Plot CSMF of the results obtained from Tariff algorithm}
\usage{
Tariff.csmf(tariff, csmf.train = c("combine", "compare", "none")[3], ...)
}
\arguments{
\item{tariff}{fitted object from \link{Tariff}}

\item{csmf.train}{option to either combine CSMF from training data (set to "combine"), or to compare with training CSMF (set to "compare"), or to ignore it (set to "none")}

\item{...}{Arguments to be passed to/from graphic function}
}
\description{
This function plots the CSMF of the fitted results.
}

