% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tariff.r
\name{Tariff}
\alias{Tariff}
\title{Replicate Tariff methods}
\usage{
Tariff(causes.train, symps.train, symps.test, use.rank = TRUE,
  nboot.rank = 1, use.sig = TRUE, nboot.sig = 100, use.top = FALSE,
  ntop = 40, causes.table = NULL)
}
\arguments{
\item{causes.train}{character vector of causes, or the column name of cause in the training data}

\item{symps.train}{N.train by S matrix, formatted as InSilicoVA input}

\item{symps.test}{N.test by S matrix, formatted as InSilicoVA input}

\item{use.rank}{logical indicator for whether using ranks instead of scores}

\item{nboot.rank}{number of re-sampling for baseline rank comparison}

\item{use.sig}{logical indicator for whether using significant Tariff only}

\item{nboot.sig}{number of re-sampling for testing significance.}

\item{use.top}{logical indicator for whether the tariff matrix should be cleaned to have only top symptoms}

\item{ntop}{number of top tariff kept for each cause}

\item{causes.table}{list of causes in the data, if different from the InterVA4 and InSilico cause list.}
}
\value{
\item{score}{matrix of score for each cause within each death}
\item{causes.train}{vector of most likely causes in training data}
\item{causes.test}{vector of most likely causes in testing data}
\item{csmf}{vector of CSMF}
\item{causes.table}{cause list used for output, i.e., list of existing causes in the training data}
}
\description{
This function implements Tariff method.
}
\examples{
data("TariffInput")
test <- TariffInput[1:100, ]
train <- TariffInput[101:200, ]
# note this is a fake dataset, so accuracy does not represent true accuracy of the algorithm!
fit <- Tariff(causes.train = "cause", symps.train = train, symps.test = test)
correct <- which(fit$causes.test == test$cause)
accuracy <- length(correct) / dim(test)[1]
}
\author{
Zehang Li, Tyler McCormick, Sam Clark

Maintainer: Zehang Li <lizehang@uw.edu>
}
\references{
Tyler H. McCormick, Zehang R. Li, Clara Calvert, Amelia C.
Crampin, Kathleen Kahn and Samuel J. Clark(2014) \emph{Probabilistic
cause-of-death assignment using verbal autopsies},
\url{http://arxiv.org/abs/1411.3042} \cr \emph{Working paper no. 147, Center
for Statistics and the Social Sciences, University of Washington}
}
\seealso{
\code{\link{Tariff.csmf}}
}
\keyword{Tariff}

