% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tariff_1.0-package.R
\docType{data}
\name{TariffInput}
\alias{TariffInput}
\title{200 records of Sample Input}
\format{200 arbitrary input records.}
\description{
This is a dataset consisting of 200 arbitrary sample input deaths in the
format of InterVA4 symptom input, with additional column at the end indicating
their cause. The causes are simulated and does not represent true causes.
}
\examples{
data(TariffInput)
}
\keyword{datasets}

