% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tariff_1.0-package.R
\docType{data}
\name{RandomVA3}
\alias{RandomVA3}
\title{400 records of Sample Input}
\format{400 arbitrary input records.}
\description{
This is a dataset consisting of 400 arbitrary sample input deaths randomly sampled from cleaned PHMRC data.
}
\examples{
data(RandomVA3)
head(RandomVA3$train)
head(RandomVA3$test)
}
\keyword{datasets}

